+++ tests/test_main.py
@@ -1,4 +1,5 @@
-""" tests """
+"""tests"""
+
 from __future__ import annotations
 
 import sys
@@ -9,6 +10,7 @@ PROJECT_DIR = Path(THISDIR).parent
 sys.path.insert(0, str(PROJECT_DIR))
 
 from tstickers.downloader import StickerDownloader
+from tstickers.convert import Backend
 
 token = ""
 for candidate in [PROJECT_DIR / "env.txt", PROJECT_DIR / "env"]:
@@ -27,22 +29,30 @@ downloader.cwd = Path(THISDIR) / "data"
 packs = [{"pack": "DonutTheDog", "len": 31}]
 
 
-def test_getPack():
+def test_getPack() -> None:
 	stickerPack = downloader.getPack(packs[0]["pack"])
 	assert stickerPack is not None
 	assert len(stickerPack["files"]) == packs[0]["len"]
 
 
-def test_downloadPack():
+def test_downloadPack() -> None:
 	stickerPack = downloader.getPack(packs[0]["pack"])
 	assert stickerPack is not None
 	downloader.downloadPack(stickerPack)
 	assert len(list(Path(f"{downloader.cwd}/donutthedog/tgs").iterdir())) == packs[0]["len"]
 
 
-def test_convertPack():
+def test_convertPack() -> None:
 	stickerPack = downloader.getPack(packs[0]["pack"])
 	assert stickerPack is not None
 	downloader.downloadPack(stickerPack)
-	downloader.convertPack(packs[0]["pack"], scale=0.05, noCache=True)
+	downloader.convertPack(packs[0]["pack"], scale=0.05, noCache=True, backend=Backend.PYRLOTTIE)
 	assert len(list(Path(f"{downloader.cwd}/donutthedog/webp").iterdir())) == packs[0]["len"]
+
+
+# def test_convertPack_slow() -> None:
+# 	stickerPack = downloader.getPack(packs[0]["pack"])
+# 	assert stickerPack is not None
+# 	downloader.downloadPack(stickerPack)
+# 	downloader.convertPack(packs[0]["pack"], scale=1, noCache=True, backend=Backend.PYRLOTTIE)
+# 	assert len(list(Path(f"{downloader.cwd}/donutthedog/webp").iterdir())) == packs[0]["len"]