+++ tstickers/__main__.py
@@ -1,4 +1,5 @@
-""" entry point for python -m tstickers """
+"""entry point for python -m tstickers."""
+
 from __future__ import annotations
 
 from . import cli