+++ tstickers/convert.py
@@ -1,37 +1,63 @@
 """Sticker convert functions used by the downloader."""
+
 from __future__ import annotations
 
-import asyncio
+import contextlib
+import multiprocessing
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
+from enum import IntEnum, auto
 from pathlib import Path
 
-import pyrlottie
 from PIL import Image
 
 
+class Backend(IntEnum):
+	UNDEFINED = -1
+	PYRLOTTIE = auto()
+	RLOTTIE_PYTHON = auto()
+
+
+def convertFunc(_x, _y, _z, _a) -> int:
+	raise RuntimeError("Backend could not be loaded")
+
+
+convertRlottiePython = convertPyRlottie = convertFunc
+
+with contextlib.suppress(ModuleNotFoundError):
+	from tstickers.convert_rlottie_python import convertAnimated as convertRlottiePython
+with contextlib.suppress(ModuleNotFoundError):
+	from tstickers.convert_pyrlottie import convertAnimated as convertPyRlottie
+
+
 def assureDirExists(root: Path, directory: Path | str) -> Path:
-	"""make the dir if not exists
+	"""Make the dir if not exists.
 
 	Args:
+	----
 		root (Path): the path of the root directory
 		directory (Path|str): the directory name
 
 	Returns:
+	-------
 		Path: the full path
+
 	"""
 	(root / directory).mkdir(parents=True, exist_ok=True)
 	return root / directory
 
 
 def convertWithPIL(inputFile: str) -> str:
-	"""Convert the webp file to png
+	"""Convert the webp file to png.
 
 	Args:
+	----
 		inputFile (str): path to input file
 
 	Returns:
+	-------
 		str: path to input file
+
 	"""
 	img = Image.open(inputFile)
 	img.save(inputFile.replace("webp", "png"))
@@ -41,15 +67,18 @@ def convertWithPIL(inputFile: str) -> str:
 	return inputFile
 
 
-def convertWebp(swd: Path, threads: int = 4) -> int:
-	"""Convert static stickers to png and gif
+def convertStatic(swd: Path, threads: int = 4) -> int:
+	"""Convert static stickers to png and gif.
 
 	Args:
+	----
 		swd (Path): the sticker working directory (downloads/packName)
 		threads (int, optional): number of threads to pass to ThreadPoolExecutor. Defaults to 4.
 
 	Returns:
+	-------
 		int: number of stickers successfully converted
+
 	"""
 	converted = 0
 	start = time.time()
@@ -69,46 +98,44 @@ def convertWebp(swd: Path, threads: int = 4) -> int:
 	return converted
 
 
-def convertTgs(swd: Path, threads: int = 4, frameSkip: int = 1, scale: float = 1) -> int:
-	"""Convert animated stickers to webp, gif and png
+def convertAnimated(
+	swd: Path,
+	threads: int = multiprocessing.cpu_count(),
+	frameSkip: int = 1,
+	scale: float = 1,
+	backend: Backend = Backend.UNDEFINED,
+) -> int:
+	"""Convert animated stickers to webp, gif and png.
 
 	Args:
+	----
 		swd (Path): the sticker working directory (downloads/packName)
-		threads (int, optional): number of threads to pass to ThreadPoolExecutor. Defaults to 4.
+		threads (int, optional): number of threads to pass to ThreadPoolExecutor. Defaults to number of cores/ logical processors.
 		frameSkip (int, optional): skip n number of frames in the interest of
 		optimisation with a quality trade-off. Defaults to 1.
 		scale (float, optional): upscale/ downscale the images produced. Intended
 		for optimisation with a quality trade-off. Defaults to 1.
 
 	Returns:
+	-------
 		int: number of stickers successfully converted
+
 	"""
-	converted = 0
+	if backend == Backend.UNDEFINED:
+		raise RuntimeError("You must specify a conversion backend")
 	start = time.time()
+	assureDirExists(swd, "apng")
 	assureDirExists(swd, "gif")
 	assureDirExists(swd, "webp")
-	doConvMultLottie = (
-		lambda fm, fs, sc: len(
-			asyncio.get_event_loop().run_until_complete(
-				pyrlottie.convMultLottie(fm, frameSkip=fs, scale=sc)
-			)
-		)
-		// 2
-	)
-	converted += doConvMultLottie(
-		fm=[
-			pyrlottie.FileMap(
-				pyrlottie.LottieFile(stckr.absolute().as_posix()),
-				{
-					stckr.absolute().as_posix().replace("tgs", "gif"),
-					stckr.absolute().as_posix().replace("tgs", "webp"),
-				},
-			)
-			for stckr in swd.glob("**/*.tgs")
-		],
-		fs=frameSkip,
-		sc=scale,
-	)
+
+	convertMap = {
+		Backend.UNDEFINED: convertFunc,
+		Backend.PYRLOTTIE: convertPyRlottie,
+		Backend.RLOTTIE_PYTHON: convertRlottiePython,
+	}
+
+	converted = convertMap[backend](swd, threads, frameSkip, scale)
+
 	end = time.time()
 	print(f"Time taken to convert {converted} stickers (tgs) - {end - start:.3f}s")
 	print()