+++ tstickers/downloader.py
@@ -13,11 +13,11 @@ from typing import Any
 from emoji import demojize
 
 from . import caching
-from .convert import assureDirExists, convertTgs, convertWebp
+from .convert import assureDirExists, convertAnimated, convertStatic, Backend
 
 
 class Sticker:
-	"""Sticker instance attributes"""
+	"""Sticker instance attributes."""
 
 	def __init__(
 		self,
@@ -25,24 +25,24 @@ class Sticker:
 		link: str = "None",
 		emoji: str = "😀",
 		fileType="webp",
-	):
+	) -> None:
 		self.name = name
 		self.link = link
 		self.emoji = emoji
 		self.fileType = fileType
 
-	def __repr__(self):
+	def __repr__(self) -> str:
 		return f"<Sticker:{self.name}>"
 
 	def emojiName(self) -> str:
-		"""get the emoji as a string"""
+		"""Get the emoji as a string."""
 		return demojize(self.emoji)[1:-1]
 
 
 class StickerDownloader:
-	"""The StickerDownloader sets up the api and makes requests"""
+	"""The StickerDownloader sets up the api and makes requests."""
 
-	def __init__(self, token, session=None, multithreading=4):
+	def __init__(self, token, session=None, multithreading=4) -> None:
 		self.threads = multithreading
 		self.token = token
 		self.cwd = assureDirExists(Path(), "downloads")
@@ -59,17 +59,21 @@ class StickerDownloader:
 			sysexit(1)
 
 	def doAPIReq(self, function: str, params: dict[Any, Any]) -> dict[Any, Any] | None:
-		"""Use the telegram api
+		"""Use the telegram api.
 
 		Args:
+		----
 			function (str): function to execute
 			params (dict[Any, Any]): function parameters
 
 		Raises:
+		------
 			RuntimeError: In the event of a failure
 
 		Returns:
+		-------
 			Optional[dict[Any, Any]]: api response
+
 		"""
 		urlParams = "?" + urllib.parse.urlencode(params)
 		res = self.session.get(f"{self.api}{function}{urlParams}")
@@ -84,34 +88,39 @@ class StickerDownloader:
 		return None
 
 	def getSticker(self, fileData: dict[Any, Any]) -> Sticker:
-		"""Get sticker info from the server
+		"""Get sticker info from the server.
 
 		Args:
+		----
 			fileData (dict[Any, Any]): sticker id
 
 		Returns:
+		-------
 			Sticker: Sticker instance
+
 		"""
 		info = self.doAPIReq("getFile", {"file_id": fileData["file_id"]})
 		if info is not None:
 			filePath = info["result"]["file_path"]
-			file = Sticker(
+			return Sticker(
 				name=filePath.split("/")[-1],
 				link=f"https://api.telegram.org/file/bot{self.token}/{filePath}",
 				emoji=fileData["emoji"],
 				fileType=filePath.split(".")[-1],
 			)
-			return file
 		return Sticker()
 
 	def getPack(self, packName: str) -> dict[str, Any] | None:
 		"""Get a list of File objects.
 
 		Args:
+		----
 			packName (str): name of the pack
 
 		Returns:
+		-------
 			dict[str, Any]: dictionary containing sticker data
+
 		"""
 		params = {"name": packName}
 		res = self.doAPIReq("getStickerSet", params)
@@ -130,22 +139,24 @@ class StickerDownloader:
 		print(f"Time taken to scrape {len(files)} stickers - {end - start:.3f}s")
 		print()
 
-		pack = {
+		return {
 			"name": res["result"]["name"].lower(),
 			"title": res["result"]["title"],
 			"files": files,
 		}
-		return pack
 
 	def downloadSticker(self, path: Path, link: str) -> int:
 		"""Download a sticker from the server.
 
 		Args:
+		----
 			path (Path): the path to write to
 			link (str): the url to the file on the server
 
 		Returns:
+		-------
 			int: path.write_bytes(res.content)
+
 		"""
 		return path.write_bytes(self.session.get(link).content)
 
@@ -153,10 +164,13 @@ class StickerDownloader:
 		"""Download a sticker pack.
 
 		Args:
+		----
 			pack (dict[str, Any]): dictionary representing a sticker pack
 
 		Returns:
+		-------
 			list[str]: list of file paths each sticker is written to
+
 		"""
 		swd = assureDirExists(self.cwd, pack["name"])
 		downloads = []
@@ -184,16 +198,20 @@ class StickerDownloader:
 		print()
 		return downloads
 
-	def convertPack(self, packName: str, frameSkip: int = 1, scale: float = 1, noCache=False):
+	def convertPack(
+		self, packName: str, frameSkip: int = 1, scale: float = 1, noCache=False, backend=Backend.UNDEFINED
+	) -> None:
 		"""Convert the webp to gif and png; tgs to gif, webp (webp_animated) and png.
 
 		Args:
+		----
 			packName (str): name of the directory to convert
 			frameSkip (int, optional): skip n number of frames in the interest of
 			optimisation with a quality trade-off. Defaults to 1.
 			scale (float, optional): upscale/ downscale the images produced. Intended
 			for optimisation with a quality trade-off. Defaults to 1.
 			noCache (bool, optional): set to true to disable cache. Defaults to False.
+
 		"""
 		if not noCache and caching.verifyConverted(packName):
 			return
@@ -207,10 +225,12 @@ class StickerDownloader:
 		print(f'Converting stickers "{packName}"...')
 
 		# tgs
-		converted = convertedTgs = convertTgs(swd, self.threads, frameSkip=frameSkip, scale=scale)
+		converted = convertedTgs = convertAnimated(
+			swd, self.threads, frameSkip=frameSkip, scale=scale, backend=backend
+		)
 
 		# webp
-		convertedWebp = convertWebp(swd, self.threads)
+		convertedWebp = convertStatic(swd, self.threads)
 		converted += convertedWebp
 
 		end = time.time()